// src/shared/api/schema.ts
export interface components {
    schemas: {
// ...
        DashboardMetricsResponse: {
            /** Totalunits */
            totalUnits: number;
            /** Totalunitscaption */
            totalUnitsCaption: string;
            occupancy: components["schemas"]["OccupancyResponse"];
            /** Occupancycaption */
            occupancyCaption: string;
            monthlyRevenue: components["schemas"]["MoneyResponse"];
            averageRate: components["schemas"]["MoneyResponse"];
            /** Activetenants */
            activeTenants: number;
            trends: components["schemas"]["DashboardMetricTrendsResponse"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}